// src/config/enums.ts
export enum Position {
  TECHNICIAN_PROGRAMMER = "technicianProgrammer",
  ENGINEER_PROGRAMMER = "engineerProgrammer",
}

export enum Level {
  INTERN = "intern",
  JUNIOR_MINUS = "juniorMinus",
  JUNIOR_EQUAL = "juniorEqual",
  JUNIOR_PLUS = "juniorPlus",
  MIDDLE_MINUS = "middleMinus",
  MIDDLE_EQUAL = "middleEqual",
  MIDDLE_PLUS = "middlePlus",
  SENIOR_MINUS = "seniorMinus",
}